// src/types/Donacion.ts
export interface Donacion {
  CreatedById: string;
  CreatedDate: string;
  detalle_medio_pago__c: string;
  Donante__c: string;
  estado__c: string;
  fecha_cobro__c: string;
  LastModifiedById: string;
  LastModifiedDate: string;
  medio_pago__c: string;
  monto__c: number;
  Name: string;
  OwnerId: string;
}